requirements.txt
annotated-types==0.7.0
anyio==4.8.0
autogen-agentchat==0.2.40
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
FLAML==2.3.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==1.26.4
openai==1.64.0
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pywin32==308
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
termcolor==2.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
